+++ src/devicon_lookup.rs
@@ -4,114 +4,11 @@ use std::ffi::OsStr;
 use std::io::{self, Write};
 use std::path::Path;
 
-lazy_static! {
-    static ref ANSI_COLOR_REGEX: Regex = Regex::new(r"\x1b\[[0-9;]*m").unwrap();
-}
+mod map;
+pub mod parsers;
 
 const DEFAULT_SYMBOL: &'static str = "";
 const ERROR_SYMBOL: &'static str = "";
-static SYMBOL_MAP: phf::Map<&'static str, &'static str> = phf_map! {
-        "ai" => "",
-        "awk" => "",
-        "bash" => "",
-        "bat" => "",
-        "bmp" => "",
-        "c++" => "",
-        "c" => "",
-        "cc" => "",
-        "clj" => "",
-        "cljc" => "",
-        "cljs" => "",
-        "coffee" => "",
-        "conf" => "",
-        "cp" => "",
-        "cpp" => "",
-        "csh" => "",
-        "css" => "",
-        "cxx" => "",
-        "d" => "",
-        "dart" => "",
-        "db" => "",
-        "diff" => "",
-        "dump" => "",
-        "edn" => "",
-        "ejs" => "",
-        "erl" => "",
-        "ex" => "",
-        "eex" => "",
-        "exs" => "",
-        "f#" => "",
-        "fish" => "",
-        "fs" => "",
-        "fsi" => "",
-        "fsscript" => "",
-        "fsx" => "",
-        "gif" => "",
-        "go" => "",
-        "h" => "",
-        "hbs" => "",
-        "hpp" => "",
-        "hrl" => "",
-        "hs" => "",
-        "htm" => "",
-        "html" => "",
-        "hxx" => "",
-        "ico" => "",
-        "ini" => "",
-        "java" => "",
-        "jl" => "",
-        "jpeg" => "",
-        "jpg" => "",
-        "js" => "",
-        "json" => "",
-        "jsx" => "",
-        "ksh" => "",
-        "less" => "",
-        "lhs" => "",
-        "lua" => "",
-        "markdown" => "",
-        "md" => "",
-        "ml" => "λ",
-        "mli" => "λ",
-        "mustache" => "",
-        "php" => "",
-        "pl" => "",
-        "pm" => "",
-        "png" => "",
-        "pp" => "",
-        "ps1" => "",
-        "psb" => "",
-        "psd" => "",
-        "py" => "",
-        "pyc" => "",
-        "pyd" => "",
-        "pyo" => "",
-        "rb" => "",
-        "rlib" => "",
-        "rmd" => "",
-        "rs" => "",
-        "rss" => "",
-        "sass" => "",
-        "scala" => "",
-        "scss" => "",
-        "sh" => "",
-        "slim" => "",
-        "sln" => "",
-        "sql" => "",
-        "styl" => "",
-        "suo" => "",
-        "t" => "",
-        "ts" => "",
-        "tsx" => "",
-        "twig" => "",
-        "vi" => "",
-        "vim" => "",
-        "vue" => "﵂",
-        "xul" => "",
-        "yaml" => "",
-        "yml" => "",
-        "zsh" => "",
-};
 
 pub type ParserResult = Result<String, &'static str>;
 pub type Parser = dyn Fn(ParserResult) -> ParserResult;
@@ -177,7 +74,7 @@ impl ParsedLine {
 
     fn symbol(&self) -> &str {
         match self.extension() {
-            Some(extension) => SYMBOL_MAP.get(extension).unwrap_or(&DEFAULT_SYMBOL),
+            Some(extension) => map::find_symbol(extension),
             None => &DEFAULT_SYMBOL,
         }
     }
@@ -188,8 +85,4 @@ impl ParsedLine {
             Err(_) => ::std::process::exit(0),
         }
     }
-}
-
-pub fn strip_color(input: ParserResult) -> ParserResult {
-    Ok(ANSI_COLOR_REGEX.replace_all(&input?, "").to_string())
 }
