+++ src/devicon_lookup/map.rs
@@ -1,0 +1,108 @@
+use phf::phf_map;
+
+static SYMBOL_MAP: phf::Map<&'static str, &'static str> = phf_map! {
+        "ai" => "",
+        "awk" => "",
+        "bash" => "",
+        "bat" => "",
+        "bmp" => "",
+        "c++" => "",
+        "c" => "",
+        "cc" => "",
+        "clj" => "",
+        "cljc" => "",
+        "cljs" => "",
+        "coffee" => "",
+        "conf" => "",
+        "cp" => "",
+        "cpp" => "",
+        "csh" => "",
+        "css" => "",
+        "cxx" => "",
+        "d" => "",
+        "dart" => "",
+        "db" => "",
+        "diff" => "",
+        "dump" => "",
+        "edn" => "",
+        "ejs" => "",
+        "erl" => "",
+        "ex" => "",
+        "eex" => "",
+        "exs" => "",
+        "f#" => "",
+        "fish" => "",
+        "fs" => "",
+        "fsi" => "",
+        "fsscript" => "",
+        "fsx" => "",
+        "gif" => "",
+        "go" => "",
+        "h" => "",
+        "hbs" => "",
+        "hpp" => "",
+        "hrl" => "",
+        "hs" => "",
+        "htm" => "",
+        "html" => "",
+        "hxx" => "",
+        "ico" => "",
+        "ini" => "",
+        "java" => "",
+        "jl" => "",
+        "jpeg" => "",
+        "jpg" => "",
+        "js" => "",
+        "json" => "",
+        "jsx" => "",
+        "ksh" => "",
+        "less" => "",
+        "lhs" => "",
+        "lua" => "",
+        "markdown" => "",
+        "md" => "",
+        "ml" => "λ",
+        "mli" => "λ",
+        "mustache" => "",
+        "php" => "",
+        "pl" => "",
+        "pm" => "",
+        "png" => "",
+        "pp" => "",
+        "ps1" => "",
+        "psb" => "",
+        "psd" => "",
+        "py" => "",
+        "pyc" => "",
+        "pyd" => "",
+        "pyo" => "",
+        "rb" => "",
+        "rlib" => "",
+        "rmd" => "",
+        "rs" => "",
+        "rss" => "",
+        "sass" => "",
+        "scala" => "",
+        "scss" => "",
+        "sh" => "",
+        "slim" => "",
+        "sln" => "",
+        "sql" => "",
+        "styl" => "",
+        "suo" => "",
+        "t" => "",
+        "ts" => "",
+        "tsx" => "",
+        "twig" => "",
+        "vi" => "",
+        "vim" => "",
+        "vue" => "﵂",
+        "xul" => "",
+        "yaml" => "",
+        "yml" => "",
+        "zsh" => "",
+};
+
+pub fn find_symbol(extenstion: &str) -> &str {
+    SYMBOL_MAP.get(extenstion).unwrap_or(&super::DEFAULT_SYMBOL)
+}
