+++ src/devicon_lookup/parsers.rs
@@ -1,0 +1,2 @@
+pub mod color;
+pub mod regex;
